+++ src/revert/function_account_email_address_verification.sql
@@ -1,6 +1,5 @@
 BEGIN;
 
-DROP FUNCTION vibetype_test.account_registration_verified(TEXT, TEXT);
 DROP FUNCTION vibetype.account_email_address_verification(UUID);
 
 COMMIT;