+++ src/verify/table_account_block.sql
@@ -6,10 +6,6 @@ SELECT id,
        created_by
 FROM vibetype.account_block WHERE FALSE;
 
-SELECT vibetype_test.index_existence(
-  ARRAY ['account_block_created_by_blocked_account_id_key']
-);
-
 ROLLBACK;
 
 BEGIN;